requirements.txt
pipecat-ai[cartesia,daily,openai,silero,gstreamer]>=0.0.58
python-dotenv~=1.0.1
asyncio
PyGObject~=3.50.0
pipecatcloud